Data handling backend
This commit adds the data backend to the document. Data is buffered in
memory unless documents are large. When taht happesn data is handled using
a temporary file.

The document type now implements the ReadSeeker interface and can be used
in all reader situations regardless if its working in memory or on disk.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>